Log In Page Corrections
Altered some functions to improve user log in validation.
Changed watchlist naming to watchlistPage.
Added validation to home page to check whether or not a user is signed in.  If no user is found, the page will redirect to the log in page.
Added functionality to the log out and watchlist buttons found on the home page.