Added Movie Class
Added a class for movies with relevant getters, setters and other functions.
Other functions only includes the "createMovie" function.
Adjusted the "createMovies" method to be inline with the newly created class.
Added a method that populates the grid with movies and their information.
Added a global click event listener that will identify if a user selects to add a movie to their watchlist.
Added some CSS styling.
Added the netflix logo image file as the url link was giving errors.